Move implementation of total_seconds to objToJSON.c; remove datetime_helper.h
